+++ src/daft-scan/src/lib.rs
@@ -45,13 +45,12 @@ pub use scan_operator::{ScanOperator, ScanOperatorRef};
 pub use scan_state::{PhysicalScanInfo, ScanState};
 pub use sharder::{Sharder, ShardingStrategy};
 pub use source_config::SourceConfig;
-pub mod shim;
 pub mod test_utils;
 
 // Re-export source module for DataSource and DataSourceTask traits.
 pub use source::{
     DataSource, DataSourceRef, DataSourceStatistics, DataSourceTask, DataSourceTaskRef,
-    DataSourceTaskStream, Precision, ReadOptions, RecordBatchStream,
+    DataSourceTaskStream, Precision, ReadOptions, RecordBatchStream, ShimSourceTask,
 };
 
 #[cfg(feature = "python")]
